% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxExtrapDemand.R
\docType{methods}
\name{approxExtrapDemand}
\alias{approxExtrapDemand}
\alias{approxExtrapDemand,LulcRasterStack-method}
\alias{approxExtrapDemand,DiscreteLulcRasterStack-method}
\title{Extrapolate land use area in time}
\usage{
approxExtrapDemand(lu, ...)

\S4method{approxExtrapDemand}{LulcRasterStack}(lu, tout, ...)

\S4method{approxExtrapDemand}{DiscreteLulcRasterStack}(lu, tout, ...)
}
\arguments{
\item{lu}{an LulcRasterStack object containing at least two maps}

\item{\dots}{additional arguments to \code{Hmisc::\link[Hmisc]{approxExtrap}}}

\item{tout}{numeric vector specifying the timesteps where interpolation is to
take place. Comparable to the \code{xout} argument of
\code{Hmisc::\link[Hmisc]{approxExtrap}}}
}
\value{
A matrix.
}
\description{
Extrapolate land use area from two or more observed land use maps to provide
a valid (although not necessarily realistic) demand scenario.
}
\details{
Many allocation routines, including the two included with \code{lulcc2},
require non-spatial estimates of land use demand for every timestep in the
study period. Some routines are coupled to complex economic models that
predict future or past land use demand based on economic considerations;
however, linear extrapolation of trends remains a useful technique.
}
\examples{

## Plum Island Ecosystems

## load observed land use maps
lu <- DiscreteLulcRasterStack(x=stack(pie[1:3]),
                              categories=c(1,2,3),
                              labels=c("Forest","Built","Other"),
                              t=c(0,6,14))

## obtain demand scenario by interpolating between observed maps
dmd <- approxExtrapDemand(lu=lu, tout=0:14)

## plot
matplot(dmd, type="l", ylab="Demand (no. of cells)", xlab="Time point",
        lty=1, col=c("Green","Red","Blue"))
legend("topleft", legend=lu@labels, col=c("Green","Red","Blue"), lty=1)

## linear extrapolation is also possible
dmd <- approxExtrapDemand(lu=lu, tout=c(0:50))

## plot
matplot(dmd, type="l", ylab="Demand (no. of cells)", xlab="Time point",
        lty=1, col=c("Green","Red","Blue"))
legend("topleft", legend=lu@labels, col=c("Green","Red","Blue"), lty=1)

}
\seealso{
\code{Hmisc::\link[Hmisc]{approxExtrap}}
}
