% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{clue}
\alias{clue}
\title{CLUE}
\usage{
clue(lu0.vals, regr, demand, elasticity, change.rule, min.elasticity,
  max.elasticity, min.change, max.change, min.value, max.value,
  max.iteration, max.difference, cell.area, ncell, ncode)
}
\arguments{
\item{lu0.vals}{matrix containing non-NA values from \code{lu0}}

\item{regr}{matrix containing...}

\item{demand}{matrix with demand for each land use category in terms of number
of cells to be allocated. The first row should be the number of cells
allocated to the initial land use map, the second row should be the number
of cells to allocate in the subsequent time point}

\item{elasticity}{Initial elasticity value. Default is 0.1}

\item{change.rule}{numeric vector specifying for each land use whether change
is allowed in either direction (0), allowed in the direction of demand only
(-1) or not allowed (1)}

\item{min.elasticity}{Minimum elasticity value. Default is 0.001}

\item{max.elasticity}{Maximum elasticity value. Default is 1.5}

\item{min.change}{numeric vector indicating for each land use the minimum
amount of change that is allowed to occur in one time step}

\item{max.change}{numeric vector indicating for each land use the maximum
amount of change that is allowed to occur in one time step}

\item{min.value}{numeric vector indicating the minimum fraction of each land
use in a given cell}

\item{max.value}{numeric vector indicating the maximum fraction of each land
use in a given cell}

\item{max.iteration}{The maximum number of iterations allowed at each time
step}

\item{max.difference}{The maximum allowable difference between demand and
allocated area}

\item{cell.area}{The area of each grid cell in the study region, which should
have the same units as the demand}

\item{ncell}{number of cells considered for change (equal to the length of
\code{lu0.vals}}

\item{ncode}{number of land use categories under consideration}

\item{\dots}{additional arguments (none)}
}
\value{
numeric vector with updated land use values.
}
\description{
Allocate land use change using the CLUE algorithm.
}
\examples{

## See lulcc2-package examples

}
