% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Model.R
\docType{class}
\name{ClueModel-class}
\alias{ClueModel-class}
\title{Class ClueModel}
\description{
An S4 class to represent inputs to the CLUE land use change model.
}
\section{Slots}{

\describe{
\item{\code{observed.lulc}}{a ContinuousLulcRasterStack object}

\item{\code{explanatory.variables}}{an ExpVarRasterStack object}

\item{\code{predictive.models}}{a PredictiveModelList object}

\item{\code{time}}{numeric vector of timesteps over which simulation will occur}

\item{\code{demand}}{matrix containing demand scenario}

\item{\code{elasticity}}{numeric}

\item{\code{change.rule}}{numeric}

\item{\code{min.elasticity}}{numeric}

\item{\code{max.elasticity}}{numeric}

\item{\code{min.value}}{numeric}

\item{\code{max.value}}{numeric}

\item{\code{min.change}}{numeric}

\item{\code{max.change}}{numeric}

\item{\code{max.iteration}}{numeric}

\item{\code{max.difference}}{numeric}

\item{\code{cell.area}}{numeric}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

