% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Model.R
\name{ClueModel}
\alias{ClueModel}
\title{Create a ClueModel object}
\usage{
ClueModel(observed.lulc, explanatory.variables, predictive.models, time, demand,
  elasticity = 0.1, change.rule, min.elasticity = 0.001,
  max.elasticity = 1.5, min.value, max.value, min.change, max.change,
  max.iteration = 1000, max.difference, cell.area)
}
\arguments{
\item{observed.lulc}{an LulcRasterStack}

\item{explanatory.variables}{an ExpVarRasterStack object}

\item{predictive.models}{a PredictiveModelList object}

\item{time}{numeric vector containing timesteps over which simulation will
occur}

\item{demand}{matrix with demand for each land use category in terms of number
of cells to be allocated. The first row should be the number of cells
allocated to the initial observed land use map (i.e. the land use map for
time 0)}

\item{elasticity}{numeric}

\item{change.rule}{numeric}

\item{min.elasticity}{numeric}

\item{max.elasticity}{numeric}

\item{min.value}{numeric}

\item{max.value}{numeric}

\item{min.change}{numeric}

\item{max.change}{numeric}

\item{max.iteration}{TODO}

\item{max.difference}{TODO}

\item{cell.area}{numeric}

\item{\dots}{additional arguments (none)}
}
\value{
A ClueModel object.
}
\description{
Methods to create a \code{ClueModel} object to supply to
\code{\link{allocate}}.
}
\examples{
## see lulcc-package examples
}
\references{
TODO
}
\seealso{
\code{\link{ClueModel-class}}, \code{\link{allocate}}
}

