% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ObsLulcRasterStack.R
\docType{methods}
\name{ContinuousObsLulcRasterStack}
\alias{ContinuousObsLulcRasterStack}
\alias{ContinuousObsLulcRasterStack,RasterStack-method}
\alias{ContinuousObsLulcRasterStack,character-method}
\title{Create an ContinuousObsLulcRasterStack object}
\usage{
ContinuousObsLulcRasterStack(x, ...)

\S4method{ContinuousObsLulcRasterStack}{RasterStack}(x, categories, labels, t)

\S4method{ContinuousObsLulcRasterStack}{character}(x, ...)
}
\arguments{
\item{x}{path (character), Raster* object or list of Raster* objects. Default
behaviour is to search for files in the working directory}

\item{categories}{numeric vector of land use categories in observed maps}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}}

\item{t}{numeric vector containing the timestep of each observed map. The
first timestep must be 0}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
An ContinuousObsLulcRasterStack object.
}
\description{
Methods to create an ContinuousObsLulcRasterStack object, which may be created
from file, an existing Raster* object or a list of Raster* objects.
}
\details{
Observed land use maps should have the same extent and resolution. The
location of non-NA cells in \code{ContinuousObsLulcRasterStack} objects
defines the region for subsequent analysis.
}
\examples{
## TODO
}
\seealso{
\code{\link{ContinuousObsLulcRasterStack-class}},
  \code{raster::\link[raster]{stack}}
}

