% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ExpVarRasterStack.R
\docType{class}
\name{ExpVarRasterStack-class}
\alias{ExpVarRasterStack-class}
\title{Class ExpVarRasterStack}
\description{
An S4 class for explanatory variables.
}
\section{Slots}{

\describe{
\item{\code{filename}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{layers}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{title}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{extent}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotated}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{rotation}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{ncols}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{nrows}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{crs}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{history}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{z}}{see \code{raster::\link[raster]{Raster-class}}}

\item{\code{index}}{data.frame TODO}
}}

