% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpVarRasterStack.R
\docType{methods}
\name{ExpVarRasterStack}
\alias{ExpVarRasterStack}
\alias{ExpVarRasterStack,RasterStack-method}
\alias{ExpVarRasterStack,character-method}
\title{Create an ExpVarRasterStack object}
\usage{
ExpVarRasterStack(x, ...)

\S4method{ExpVarRasterStack}{character}(x, ...)

\S4method{ExpVarRasterStack}{RasterStack}(x, index, ...)
}
\arguments{
\item{x}{Raster* object}

\item{index}{data.frame}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
An ExpVarRasterStack object.
}
\description{
Methods to ...
}
\examples{
\dontrun{

## Plum Island Ecosystems

ef <- ExpVarRasterStack(x=pie, pattern="ef")

## Sibuyan
ef <- ExpVarRasterStack(x=sibuyan$maps, pattern="ef")

}
}
\seealso{
\code{raster::\link[raster]{stack}}
}

