% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpVarRasterList.R
\docType{methods}
\name{ExpVarRasterStack}
\alias{ExpVarRasterStack}
\alias{ExpVarRasterStack,RasterStack,character-method}
\alias{ExpVarRasterStack,RasterStack-method}
\alias{ExpVarRasterStack,character,character-method}
\alias{ExpVarRasterStack,character-method}
\title{Create an ExpVarRasterStack object}
\usage{
ExpVarRasterStack(x, ...)

\S4method{ExpVarRasterStack}{character}(x, ...)

\S4method{ExpVarRasterStack}{RasterStack}(x, index, ...)
}
\arguments{
\item{x}{path (character) to directory containing observed land use maps,
a Raster* object or a list of Raster* objects}

\item{index}{data.frame}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
An ExpVarRasterStack object.
}
\description{
Methods to...
}
\examples{
## TODO
}
\seealso{
\code{raster::\link[raster]{stack}}
}

