% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/index.R
\docType{methods}
\name{Extract by index}
\alias{Extract by index}
\alias{[[,ContinuousObsLulcRasterStack,ANY,ANY-method}
\alias{[[,DiscreteObsLulcRasterStack,ANY,ANY-method}
\alias{[[,PerformanceList,ANY,ANY-method}
\alias{[[,PredictionList,ANY,ANY-method}
\alias{[[,PredictiveModelList,ANY,ANY-method}
\title{Extract by index}
\usage{
\S4method{[[}{DiscreteObsLulcRasterStack,ANY,ANY}(x, i, j, ...)

\S4method{[[}{PerformanceList,ANY,ANY}(x, i, j, ...)

\S4method{[[}{PredictionList,ANY,ANY}(x, i, j, ...)

\S4method{[[}{PredictiveModelList,ANY,ANY}(x, i, j, ...)

\S4method{[[}{ContinuousObsLulcRasterStack,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{TODO}

\item{i}{layer number (if 'x' inherits from a RasterStack) or list index (if
'x' stores data as a list)}

\item{j}{numeric (not used)}

\item{...}{additional arguments (none)}
}
\description{
\code{object[[i]]} can be used to extract individual objects from container
classes such as \code{ExpVarRasterStack}, \code{PredictiveModelList}, \code{PredictionList} and
\code{PerformanceList}.
}
\examples{
## Plum Island Ecosystems

## Load observed land use maps
obs <- ObsLulcRasterStack(x=pie,
                   pattern="lu",
                   categories=c(1,2,3),
                   labels=c("forest","built","other"),
                   t=c(0,6,14))

summary(obs[[1]])
summary(obs[[1:2]])
}

