% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictiveModelInputData.R
\name{getPredictiveModelInputData}
\alias{getPredictiveModelInputData}
\title{Extract data to fit predictive models}
\usage{
getPredictiveModelInputData(lu, ef, cells, ...)
}
\arguments{
\item{lu}{an LulcRasterStack object}

\item{ef}{an ExpVarRasterStack object}

\item{cells}{index of cells to be extracted, which may be a
\code{SpatialPoints*} object or a numeric vector representing cell numbers
(see \code{raster::\link[raster]{extract}})}

\item{...}{additional arguments to \link{as.data.frame}}
}
\value{
A data.frame.
}
\description{
Extract a data.frame containing variables required for fitting predictive
models. Column names correspond to the names of lu and ef.
}
\examples{

\dontrun{

## Plum Island Ecosystems

lu <- DiscreteLulcRasterStack(x=stack(pie[1:3]),
                              categories=c(1,2,3),
                              labels=c("Forest","Built","Other"),
                              t=c(0,6,14))

idx <- data.frame(var=paste("ef_", formatC(1:3, width=3, flag=0)),
                  yr=rep(0,3),
                  dynamic=rep(FALSE,3))

ef <- ExpVarRasterStack(x=stack(pie[4:6]), index=idx)

part <- partition(x=lu, size=0.1, spatial=TRUE, t=0)

train.data <- getPredictiveModelInputData(lu=lu,
                                          ef=ef,
                                          cells=part[["train"]],
                                          t=0)
dim(train.data)
names(train.data)

}

}
\seealso{
\code{\link[base]{as.data.frame}}, \code{\link{LulcRasterStack}},
\code{\link{ExpVarRasterStack}}, \code{\link{partition}}
}
