% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LulcRasterStack.R
\docType{methods}
\name{LulcRasterStack}
\alias{ContinuousLulcRasterStack}
\alias{ContinuousLulcRasterStack,Raster-method}
\alias{ContinuousLulcRasterStack,RasterStack-method}
\alias{DiscreteLulcRasterStack}
\alias{DiscreteLulcRasterStack,Raster-method}
\alias{DiscreteLulcRasterStack,RasterStack-method}
\alias{LulcRasterStack}
\title{Create an LulcRasterStack object}
\usage{
DiscreteLulcRasterStack(x, ...)

\S4method{DiscreteLulcRasterStack}{Raster}(x, ...)

\S4method{DiscreteLulcRasterStack}{RasterStack}(x, pattern, categories, labels,
  t)

ContinuousLulcRasterStack(x, ...)

\S4method{ContinuousLulcRasterStack}{Raster}(x, ...)

\S4method{ContinuousLulcRasterStack}{RasterStack}(x, categories, labels, t)
}
\arguments{
\item{x}{path (character), Raster* object or list of Raster* objects. Default
behaviour is to search for files in the working directory}

\item{pattern}{regular expression (character). Only filenames (if \code{x} is
a path) or Raster* objects (if \code{x} is a list) matching the regular
expression will be returned. See \cr
\code{raster::\link[raster]{raster}} for more information about supported filetypes}

\item{categories}{numeric vector of land use categories in observed maps}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}}

\item{t}{numeric vector containing the timestep of each observed map. The
first timestep must be 0}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}
}
\value{
An LulcRasterStack object.
}
\description{
Methods to create an LulcRasterStack object, which may be created from file, an
existing Raster* object or a list of Raster* objects.
}
\details{
Observed land use maps should have the same extent and resolution. The
location of non-NA cells in \code{LulcRasterStack} objects defines the region for
subsequent analysis.
}
\examples{
\dontrun{

## Plum Island Ecosystems
obs <- LulcRasterStack(x=pie,
                   pattern="lu",
                   categories=c(1,2,3),
                   labels=c("forest","built","other"),
                   t=c(0,6,14))

## Sibuyan Island
obs <- LulcRasterStack(x=sibuyan$maps,
                   pattern="lu",
                   categories=c(1,2,3,4,5),
                   labels=c("forest","coconut","grass","rice","other"),
                   t=c(0,14))

}
}
\seealso{
\code{\link{LulcRasterStack-class}}, \code{raster::\link[raster]{stack}}
}

