% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PredictiveModelList.R
\name{Model fitting}
\alias{Model fitting}
\alias{glmModels}
\alias{lmModels}
\alias{randomForestModels}
\title{Fit predictive models}
\usage{
lmModels(formula, ..., obs, categories = NA, labels = NA)

glmModels(formula, family = binomial, ..., obs, categories = NA,
  labels = NA)

randomForestModels(formula, ..., obs, categories = NA, labels = NA)
}
\arguments{
\item{formula}{list containing formula objects}

\item{...}{additional arguments to specific functions}

\item{obs}{TODO}

\item{categories}{TODO}

\item{labels}{TODO}

\item{family}{see \code{\link[stats]{glm}}. Default is 'binomial'. Only used by
\code{glmModels}}
}
\value{
A PredictiveModelList object
}
\description{
These functions fit parametric and non-parametric models to data.
}
\examples{
## see lulcc-package examples
}
\seealso{
\code{\link[stats]{glm}},
  \code{randomForest::\link[randomForest]{randomForest}}
}

