% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Model.R
\docType{class}
\name{OrderedModel-class}
\alias{OrderedModel-class}
\title{Class OrderedModel}
\description{
An S4 class to represent inputs to the Ordered allocation procedure
}
\section{Slots}{

\describe{
\item{\code{observed.lulc}}{an LulcRasterStack object}

\item{\code{explanatory.variables}}{an ExpVarRasterStack object}

\item{\code{predictive.models}}{a PredictiveModelList object}

\item{\code{time}}{numeric vector of timesteps over which simulation will occur}

\item{\code{demand}}{matrix containing demand scenario}

\item{\code{history}}{RasterLayer showing land use history or NULL}

\item{\code{mask}}{RasterLayer showing masked areas or NULL}

\item{\code{neighbourhood}}{NeighbRasterStack object or NULL}

\item{\code{transition.rules}}{matrix with land use change decision rules}

\item{\code{neighbourhood.rules}}{numeric with neighbourhood decision rules}

\item{\code{order}}{numeric vector of land use categories in the order that change
should be allocated}

\item{\code{categories}}{numeric vector of land use categories}

\item{\code{labels}}{character vector corresponding to \code{categories}}
}}

