% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{OrderedModel}
\alias{OrderedModel}
\title{Create a OrderedModel object}
\usage{
OrderedModel(observed.lulc, explanatory.variables, predictive.models, time,
  demand, history = NULL, mask = NULL, neighbourhood = NULL,
  transition.rules, neighbourhood.rules = NULL, order, ...)
}
\arguments{
\item{observed.lulc}{an LulcRasterStack}

\item{explanatory.variables}{an ExpVarRasterStack object}

\item{predictive.models}{a PredictiveModelList object}

\item{time}{numeric vector containing timesteps over which simulation will
occur}

\item{demand}{matrix with demand for each land use category in terms of number
of cells to be allocated. The first row should be the number of cells
allocated to the initial observed land use map (i.e. the land use map for
time 0)}

\item{history}{RasterLayer containing land use history (values represent the
number of years the cell has contained the current land use category)}

\item{mask}{RasterLayer containing binary values where 0 indicates cells
that are not allowed to change}

\item{neighbourhood}{an object of class NeighbRasterStack}

\item{transition.rules}{matrix with land use change decision rules}

\item{neighbourhood.rules}{numeric with neighbourhood decision rules}

\item{order}{numeric vector of land use categories in the order that change
should be allocated}

\item{\dots}{additional arguments (none)}
}
\value{
A OrderedModel object.
}
\description{
Methods to create a \code{OrderedModel} object to supply to
\code{\link{allocate}}.
}
\examples{

## see lulcc2-package examples
}
\references{
Fuchs, R., Herold, M., Verburg, P.H., and Clevers, J.G.P.W. (2013). A
high-resolution and harmonized model approach for reconstructing and analysing
historic land changes in Europe, Biogeosciences, 10:1543-1559.
}
\seealso{
\code{\link{OrderedModel-class}}, \code{\link{allocate}}
}
