% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\docType{methods}
\name{partition}
\alias{partition}
\alias{partition,RasterLayer-method}
\alias{partition,DiscreteLulcRasterStack-method}
\alias{partition,ContinuousLulcRasterStack-method}
\title{Partition raster data}
\usage{
partition(x, ...)

\S4method{partition}{RasterLayer}(x, size = 0.5, spatial = TRUE, ...)

\S4method{partition}{DiscreteLulcRasterStack}(x, size = 0.5,
  spatial = TRUE, t, ...)

\S4method{partition}{ContinuousLulcRasterStack}(x, size = 0.5,
  spatial = TRUE, ...)
}
\arguments{
\item{x}{RasterLayer with categorical data}

\item{\dots}{additional arguments (none)}

\item{size}{numeric value between zero and one indicating the proportion of
non-NA cells that should be included in the training partition. Default is
0.5, which results in equally sized partitions}

\item{spatial}{logical. If TRUE, the function returns a SpatialPoints object
with the coordinates of cells in each partition. If FALSE, the cell numbers
are returned}

\item{t}{numeric corresponding to one of the time points for which a land use
map is available.}
}
\value{
A list containing the following components:
\describe{
  \item{\code{train}}{a SpatialPoints object or numeric vector indicating the
  cells in the training partition}
  \item{\code{test}}{a SpatialPoints object or numeric vector indicating the
  cells in the testing partition}
  \item{\code{all}}{a SpatialPoints object or numeric vector indicating all
  non-NA cells in the study region}
}
}
\description{
Divide a categorical raster map into training and testing partitions.
A wrapper function for \cr
\code{caret::\link[caret]{createDataPartition}} (Kuhn, 2008).
}
\examples{

\dontrun{

## Plum Island Ecosystems

lu <- DiscreteLulcRasterStack(x=stack(pie[1:3]),
                              categories=c(1,2,3),
                              labels=c("Forest","Built","Other"),
                              t=c(0,6,14))

part <- partition(x=lu, size=0.05, spatial=TRUE, t=0)

plot(lu[[1]])
points(part[["train"]])

}
}
\references{
Kuhn, M. (2008). Building predictive models in R using the caret
package. Journal of Statistical Software, 28(5), 1-26.
}
\seealso{
\code{caret::\link[caret]{createDataPartition}}
}
