% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictiveModelList.R
\name{PredictiveModelList}
\alias{PredictiveModelList}
\title{Create PredictiveModelList object}
\usage{
PredictiveModelList(models, categories, labels, ...)
}
\arguments{
\item{models}{list containing predictive models}

\item{categories}{numeric vector of land use categories in observed maps}

\item{labels}{character vector with labels corresponding to \code{categories}}

\item{...}{additional arguments (none)}
}
\value{
A PredictiveModelList object
}
\description{
Crete an object of class PredictiveModelList.
}
\examples{

## see lulcc2-package examples
}
