% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\docType{methods}
\name{resample,ExpVarRasterStack,Raster-method}
\alias{resample,ExpVarRasterStack,Raster-method}
\alias{resample,list,Raster-method}
\title{Resample maps in ExpVarRasterStack object or list}
\usage{
\S4method{resample}{ExpVarRasterStack,Raster}(x, y, method = "ngb", ...)

\S4method{resample}{list,Raster}(x, y, method = "ngb", ...)
}
\arguments{
\item{x}{an ExpVarRasterStack object or list of Raster* maps to be resampled}

\item{y}{Raster* object with parameters that \code{x} should be resampled to}

\item{method}{method used to compute values for the new RasterLayer, should be
\code{"bilinear"} for bilinear interpolation, or \code{"ngb"} for nearest
neighbour}

\item{\dots}{additional arguments to \code{raster::\link[raster]{resample}}}
}
\value{
An ExpVarRasterStack object or list, depending on \code{x}.
}
\description{
A wrapper function for \code{raster::\link[raster]{resample}} to resample
raster objects in an ExpVarRasterStack object or list.
}
\examples{

## see lulcc2-examples
}
\seealso{
\code{\link{ExpVarRasterStack}}, \code{raster::\link[raster]{resample}}
}
