% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,PredictiveModelList-method}
\alias{show,PredictiveModelList-method}
\alias{show,PredictionList-method}
\alias{show,PerformanceList-method}
\alias{show,Model-method}
\alias{show,ThreeMapComparison-method}
\title{Show objects}
\usage{
\S4method{show}{PredictiveModelList}(object)

\S4method{show}{PredictionList}(object)

\S4method{show}{PerformanceList}(object)

\S4method{show}{Model}(object)

\S4method{show}{ThreeMapComparison}(object)
}
\arguments{
\item{object}{an object belonging to one of the classes in \code{lulcc2}}
}
\value{
Null
}
\description{
Show lulcc2 objects
}
