% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{subset,PredictiveModelList-method}
\alias{subset,PredictiveModelList-method}
\alias{subset,PerformanceList-method}
\alias{subset,PredictionList-method}
\title{Subset objects}
\usage{
\S4method{subset}{PredictiveModelList}(x, subset, drop = FALSE, ...)

\S4method{subset}{PerformanceList}(x, subset, ...)

\S4method{subset}{PredictionList}(x, subset, ...)
}
\arguments{
\item{x}{an object of class \code{ExpVarRasterStack},
\code{PredictiveModelList}, \code{PredictionList} or \code{PerformanceList}}

\item{subset}{integer or character indicating the objects to be extracted}

\item{drop}{logical}

\item{...}{additional arguments (none)}
}
\value{
Subsetted object, possibly simplified
}
\description{
Extract a subset of objects from container classes such as
\code{ExpVarRasterStack}, \code{PredictiveModelList}, \code{PredictionList} and
\code{PerformanceList}.
}
\examples{

## Sibuyan Island

lu <- DiscreteLulcRasterStack(x=stack(sibuyan$maps[1:2]),
                              categories=c(1,2,3,4,5),
                              labels=c("forest","coconut","grass","rice","other"),
                              t=c(0,14))


summary(lu)
lu <- subset(lu, subset=names(lu)[1])
summary(lu)

## load explanatory variables
idx <- data.frame(var=paste("ef_", formatC(1:13, width=3, flag=0)),
                  yr=rep(0,13),
                  dynamic=rep(FALSE,13))

ef <- ExpVarRasterStack(x=stack(sibuyan$maps[3:15]), index=idx)

summary(ef)
ef <- subset(ef, subset=1:5)
summary(ef)

}
