% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,LulcRasterStack-method}
\alias{summary,ExpVarRasterStack-method}
\alias{summary,Model-method}
\title{Summary}
\usage{
summary(object, ...)

\S4method{summary}{LulcRasterStack}(object, ...)

\S4method{summary}{ExpVarRasterStack}(object, ...)

\S4method{summary}{Model}(object, ...)
}
\arguments{
\item{object}{an object belonging to one of the classes in \code{lulcc2}}

\item{...}{additional arguments (none)}
}
\value{
A matrix, data.frame or list
}
\description{
Summarise lulcc2 objects
}
