% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeMapComparison.R
\name{threeDimensionTable}
\alias{threeDimensionTable}
\title{Calculate three dimensional comparison table}
\usage{
threeDimensionTable(x.list, x1.list, y1.list, wt.vals, categories, ...)
}
\arguments{
\item{x.list}{list}

\item{x1.list}{list}

\item{y1.list}{list}

\item{wt.vals}{list}

\item{categories}{numeric}

\item{\dots}{additional arguments (none)}
}
\value{
A list.
}
\description{
Calculate three dimensional comparison table
}
