% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\docType{methods}
\name{total}
\alias{total}
\alias{total,DiscreteLulcRasterStack-method}
\alias{total,ContinuousLulcRasterStack-method}
\alias{total,Raster-method}
\title{Total number of cells in a Raster* object}
\usage{
total(x, ...)

\S4method{total}{DiscreteLulcRasterStack}(x, categories, ...)

\S4method{total}{ContinuousLulcRasterStack}(x, categories, ...)

\S4method{total}{Raster}(x, categories, ...)
}
\arguments{
\item{x}{Raster* object}

\item{\dots}{additional arguments (none)}

\item{categories}{numeric vector containing land use categories. Only cells
belonging to these categories will be counted}
}
\value{
A list containing the following components:
\describe{
  \item{\code{total}}{a matrix containing the total number of cells belonging
    to each category. Rows represent layers in the input Raster* object}
  \item{\code{categories}}{the categories included in the calculation}
}
}
\description{
Count the area or number of cells belonging to each category in a Raster*
object.
}
\details{
If x is a DiscreteLulcRasterStack object this function returns the number of
cells belonging to each category. If x is a ContinuousLulcRasterStack object
the function returns the sum of the fractions of the various land use
categories.
}
\examples{

## Sibuyan Island

lu <- DiscreteLulcRasterStack(x=stack(sibuyan$maps[1:2]),
                              categories=c(1,2,3,4,5),
                              labels=c("forest","coconut","grass","rice","other"),
                              t=c(0,14))

total(x=lu)
total(x=lu[[1]])
total(x=lu[[2]])

}
