% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.data.frame.R
\docType{methods}
\name{updateDataFrame}
\alias{updateDataFrame}
\alias{updateDataFrame,ExpVarRasterStack-method}
\title{Update data.frame}
\usage{
updateDataFrame(x, ...)

\S4method{updateDataFrame}{ExpVarRasterStack}(x, y, cells, time, ...)
}
\arguments{
\item{x}{ExpVarRasterStack object}

\item{y}{data.frame to update}

\item{cells}{TODO}

\item{time}{TODO}

\item{\dots}{TODO}
}
\value{
data.frame
}
\description{
TODO
}
\examples{
## TODO
}

