% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{ClueModel}
\alias{ClueModel}
\title{Create a ClueModel object}
\usage{
ClueModel(observed.lulc, explanatory.variables, predictive.models, time, demand,
  elasticity = 0.1, change.rule, min.elasticity = 0.001,
  max.elasticity = 1.5, min.value, max.value, min.change, max.change,
  max.iteration = 1000, max.difference, cell.area)
}
\arguments{
\item{observed.lulc}{an LulcRasterStack}

\item{explanatory.variables}{an ExpVarRasterStack object}

\item{predictive.models}{a PredictiveModelList object}

\item{time}{numeric vector containing timesteps over which simulation will
occur}

\item{demand}{matrix with demand for each land use category in terms of number
of cells to be allocated. The first row should be the number of cells
allocated to the initial observed land use map (i.e. the land use map for
time 0)}

\item{elasticity}{Initial elasticity value. Default is 0.1}

\item{change.rule}{numeric vector specifying for each land use whether change
is allowed in either direction (0), allowed in the direction of demand only
(-1) or not allowed (1)}

\item{min.elasticity}{Minimum elasticity value. Default is 0.001}

\item{max.elasticity}{Maximum elasticity value. Default is 1.5}

\item{min.value}{numeric vector indicating the minimum fraction of each land
use in a given cell}

\item{max.value}{numeric vector indicating the maximum fraction of each land
use in a given cell}

\item{min.change}{numeric vector indicating for each land use the minimum
amount of change that is allowed to occur in one time step}

\item{max.change}{numeric vector indicating for each land use the maximum
amount of change that is allowed to occur in one time step}

\item{max.iteration}{The maximum number of iterations allowed at each time
step}

\item{max.difference}{The maximum allowable difference between demand and
allocated area}

\item{cell.area}{The area of each grid cell in the study region, which should
have the same units as the demand}

\item{\dots}{additional arguments (none)}
}
\value{
A ClueModel object.
}
\description{
Methods to create a \code{ClueModel} object to supply to
\code{\link{allocate}}.
}
\examples{

## see lulcc2-package examples
}
\references{
Veldkamp, A., & Fresco, L. O. (1996). CLUE-CR: an integrated multi-scale model
to simulate land use change scenarios in Costa Rica. Ecological modelling,
91(1), 231-248.

Verburg, P.H., & Bouma, J. (1999). Land use change under conditions of high
population pressure: the case of Java. Global environmental change, 9(4),
303-312.
}
\seealso{
\code{\link{ClueModel-class}}, \code{\link{allocate}}
}
