% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\docType{methods}
\name{Extract by index}
\alias{Extract by index}
\alias{[[,DiscreteLulcRasterStack,ANY,ANY-method}
\alias{[[,PerformanceList,ANY,ANY-method}
\alias{[[,PredictionList,ANY,ANY-method}
\alias{[[,PredictiveModelList,ANY,ANY-method}
\alias{[,PredictiveModelList,ANY,ANY-method}
\alias{[[,ContinuousLulcRasterStack,ANY,ANY-method}
\title{Extract by index}
\usage{
\S4method{[[}{DiscreteLulcRasterStack,ANY,ANY}(x, i, j, ...)

\S4method{[[}{PerformanceList,ANY,ANY}(x, i, j, ...)

\S4method{[[}{PredictionList,ANY,ANY}(x, i, j, ...)

\S4method{[[}{PredictiveModelList,ANY,ANY}(x, i, j, ...)

\S4method{[}{PredictiveModelList,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ContinuousLulcRasterStack,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{An object of class DiscreteLulcRasterStack,
ContinuousLulcRasterStack, PredictionList, PerformanceList,
PredictiveModelList}

\item{i}{layer number (if 'x' inherits from a RasterStack) or list index (if
'x' stores data as a list)}

\item{j}{numeric (not used)}

\item{...}{additional arguments (none)}

\item{drop}{logical. If TRUE the result is coerced to the lowest possible
dimension}
}
\description{
\code{object[[i]]} can be used to extract individual objects from container
classes such as \code{ExpVarRasterStack}, \code{PredictiveModelList},
\code{PredictionList} and \code{PerformanceList}.
}
\examples{

## Plum Island Ecosystems

lu <- DiscreteLulcRasterStack(x=stack(pie[1:3]),
                              categories=c(1,2,3),
                              labels=c("Forest","Built","Other"),
                              t=c(0,6,14))

summary(lu[[1]])
summary(lu[[1:2]])

## Also see lulcc2-package
}
