% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LulcRasterStack.R
\docType{methods}
\name{LulcRasterStack}
\alias{LulcRasterStack}
\alias{DiscreteLulcRasterStack}
\alias{DiscreteLulcRasterStack,Raster-method}
\alias{DiscreteLulcRasterStack,RasterStack-method}
\alias{ContinuousLulcRasterStack}
\alias{ContinuousLulcRasterStack,Raster-method}
\alias{ContinuousLulcRasterStack,RasterStack-method}
\title{Create an LulcRasterStack object}
\usage{
DiscreteLulcRasterStack(x, ...)

\S4method{DiscreteLulcRasterStack}{Raster}(x, ...)

\S4method{DiscreteLulcRasterStack}{RasterStack}(x, categories, labels, t)

ContinuousLulcRasterStack(x, ...)

\S4method{ContinuousLulcRasterStack}{Raster}(x, ...)

\S4method{ContinuousLulcRasterStack}{RasterStack}(x, categories, labels, t)
}
\arguments{
\item{x}{path (character), Raster* object or list of Raster* objects. Default
behaviour is to search for files in the working directory}

\item{\dots}{additional arguments to \code{raster::\link[raster]{stack}}}

\item{categories}{numeric vector of land use categories in observed maps}

\item{labels}{character vector (optional) with labels corresponding to
\code{categories}}

\item{t}{numeric vector containing the timestep of each observed map. The 
first timestep must be 0}
}
\value{
An LulcRasterStack object.
}
\description{
Methods to create an LulcRasterStack object, which may be created from file
or an existing Raster* object
}
\details{
Observed land use maps should have the same extent and resolution and have
the same non-NA cells. The location of non-NA cells in \code{LulcRasterStack}
objects defines the region for subsequent analysis.
}
\examples{

## Plum Island Ecosystems 

lu <- DiscreteLulcRasterStack(x=stack(pie[1:3]),
                              categories=c(1,2,3),
                              labels=c("Forest","Built","Other"),
                              t=c(0,6,14))
lu

## Sibuyan Island
lu <- DiscreteLulcRasterStack(x=stack(sibuyan$maps[1:2]),
                              categories=c(1,2,3,4,5),
                              labels=c("forest","coconut","grass","rice","other"),
                              t=c(0,14))

}
\seealso{
\code{\link{LulcRasterStack-class}}, \code{raster::\link[raster]{stack}}
}
