% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\docType{methods}
\name{updateDataFrame}
\alias{updateDataFrame}
\alias{updateDataFrame,ExpVarRasterStack-method}
\title{Update data frame}
\usage{
updateDataFrame(x, ...)

\S4method{updateDataFrame}{ExpVarRasterStack}(x, y, cells, time, ...)
}
\arguments{
\item{x}{ExpVarRasterStack object}

\item{\dots}{additional arguments (none)}

\item{y}{data.frame to update}

\item{cells}{index of cells to be extracted and added to the data frame (see
\code{\link{as.data.frame}})}

\item{time}{numeric indicating the time for which the data frame should be
updated}
}
\value{
data.frame
}
\description{
Function to update a data frame holding model variables with values of dynamic
covariables for a new time point. This function is used internally by
allocation routines.
}
